Perform dealiasing in Namer, not Typer
This simple change makes a big difference in Typer. All types and effects can
now safely be assumed to be dealiased. That is, we do not need ad-hoc interleaving
of substitution and dealiasing. It also means that it is easily possible to compute
the canonical ordering of capabilities for a given set of effects.