Implement monadic and callcc backend for Scheme again
There is a bug with anf.md

The problem is that toplevel declarations have to be run in a mounted
meta-continuation. So either, they need to be wrapped in:

`(run (thunk EXPR))`

or the whole user program has to be wrapped in `run`.

However, the library needs to be prefixed outside of the run, of course
to be in scope.