Directly forward execution output to stdout
Issue #336 is caused by the Process being started using `.!!` which:
```
Starts the process represented by this builder, blocks until it exits, and returns the output as a String. Standard error is sent to the console. If the exit code is non-zero, an exception is thrown.
```

This uses `.!<` instead, which:
```
Starts the process represented by this builder, blocks until it exits, and returns the exit code. Standard output and error are sent to the console. The newly started process reads from standard input of the current process.
```

Note: This should also allow reading from stdin now (but I have not
tested this yet).