Hotfix: box/unbox function arguments to expected type first
This will:
- Potentially incur double coercions for arguments
- Probably not work in some cases, as we do not track the return type

But this does fix the bugs we are currently running into.

Note: In the future, a more complete rewrite is needed, as it was
written with parametric polymorphism in mind and is now used for
subtyping. This means the assumption that casts *only* need to happen
at instantiation sites (i.e. call sites) is broken.