Add more GCC flags to CI builds (#656)
Within the last few weeks we've repeatedly had compilation warnings in
the LLVM backend compiled with the master branch. The warnings could
have been caught earlier by running Effekt with the `--debug` flag.

To enforce this, the CI now runs with `--debug` by default (in addition
to `--valgrind`) and `--debug` adds the additional `-Werror` flag --
thus failing the CI if the compilation produces warnings.