Add immutable ordered finite 'map' and 'set' to the stdlib (#561)
I've been sitting on this code for almost ~~9~~ _13_ months now (started
on Xmas 2023), it's about time we finalise it.
This PR adds immutable ordered maps and sets based on weight balanced
trees into the standard library.

The `map` and `set` now work on all main backends: JS, LLVM, Chez.

Progress ~~is **blocked**~~ was blocked on:
1. ~~very annoying inliner bug(s?) [#733]~~ _resolved_ 🥳 
2. ~~missing comparisons on strings on LLVM [#748]~~ _resolved_ 🎉