Prevent stack overflows by using larger stacks (#853)
Closes #847. Error occurred in #828, #841, #851, #852.

Unfortunately I don't think we can set the stack size in `build.sbt`
unless we use a forked JVM, so we have to set it manually using an
argument to SBT (or an environment variable).

The problem mainly comes from our abundant use of non-tail recursion, we
could refactor this in the future. We should also decide on a specific
stack size, I've set it to 1G for now which is quite large.