Extend array stdlib (#839)
closes #804

This PR aims to extend the stdlib of `array`. Most of the added
functions are in-place compared to their counterparts for `list`.

Feel free to bike-shed. I am looking forward to your feedback.

## TODO

- [x] Add complete test suit for all functions
- [x] There is `indexOf`, `lastIndexOf` and `find`, probably not all of
them are needed.
- [x] It would be nice if someone could check the implementation
`c_array_fill` as I am not too familiar with LLVM and the memory
management.
- [x] ~~Decide on naming convention: ruby- vs scala-style (e.g. `map!`
vs `mapped`)?~~ Change everything to Ruby style
- [x] Measure performance of the extern definition of `fill` compared to
a definition in effekt