Name join points introduced when compiling clauses (#848)
Quick drive-by from #841 that deserves to be considered separately.

Instead of naming each join point created for the RHS of a clause as
`b_k_123`, we can give a slightly more specific name based on the origin
of the clause.

Feel free to bikeshed the names chosen or close this PR altogether if
it's useless.
FWIW, it helps me ever-so-slightly in the `b_k${index of the case in the
source}_123` case when debugging.