Fix canonical ordering by using the ordering at the declaration site (#859)
This tries to fix issue #849 by NOT canonicalizing AFTER unification.

Instead, we canonicalize FIRST on the function type and then MAINTAIN
the order at the callsite.

This requires us to NOT go into type `Effect`, which is a set and thus
removes duplicates and destroys the ordering...

Sadly, it is very difficult to write a test for this :(