CI: Extract reusable 'test' and 'build' actions (#888)
I was sitting on this idea for a while: the CI is somewhat difficult to
navigate since we keep building the env anew every single time...

GitHub Actions allows us to make reusable actions which we can use like
subroutines for the actual workflows.
Here, we have:
- `run-effekt-tests` which just runs tests (we can choose whether full
tests or partial ones or if it should retry)
- `setup-effekt` which prepares the devenv

See docs
[here](https://docs.github.com/en/actions/sharing-automations/avoiding-duplication).

Now it's much easier to, say, change the LLVM version in _all_ CI
actions! :^)

---

No idea if this works, most of this was vibe-coded with Sonnet 3.7 and I
have no way of testing this locally anyways... 🤷‍♂️ 😎