Fix: make type pretty printer less confusing for boxed types (#906)
Fixes #659, but not #667. I propose deferring fixing #667 to a later
point in time, as it requires more design.

What this PR does, for now, is to parenthesize boxed types when
occurring in parameter or return type positions. For example, `(Int =>
Int at {}) => Int` instead of `Int => Int at {} => Int`, which, in my
opinion, is utterly unreadable.