Simplify core.Renamer to produce smaller symbols (#938)
General idea: instead of building
`name_123_109238_190238_109283019283_1092830192838019283` chains, what
if we use the automagic `fresh.next` to get just a darn fresh number?
This will limit the identifiers to stuff like `b_k2_123_10928310283`,
which is--at least to me--a pretty big improvement.

We use the full capabilities of a `Symbol`: it has a `name` (which is
really a `String`) and a fresh `id` that gets assigned lazily. Instead
of changing the name to be `${oldName}_${oldId}` and then putting a
fresh `id` on top of that (to get `${oldName}_${oldId}_${newId}`), we
just make a new `Symbol` with the same name, which forces a new,
_fresher_, `id`.

---------

Co-authored-by: Marcial Gaißert <github@gaisseml.de>