Make 'core.Renamer' quicker by simplifying scope handling (#940)
I don't know what the nested list is for, but we don't seem to use it
anywhere. Therefore I'm removing the nesting and using mutable hashmaps:
it's on a hot path!

Renamer then goes from 2.4-3.4s to 0.6-0.4s :) [measured a few times via
profiling]
(for reference, before #938, it was somewhere between 3 and 4 seconds)