Cache phases Parser until Typer (#920)
Previously, our annotations relied on the object identity of the syntax
tree that directly results from parsing.
However, some phases, in particular the `BoxUnboxInference` phase,
rewrite the tree which changes the object identity of nodes.
This led to issues such as #897 and #549.
Finishing what we started at the end of the Hackathon last friday, this
PR now caches the phases `Parser` until `Typer` rather than just `Typer`
and uses the resulting tree for `getAST` queries.
This is probably not the final solution for this problem, but it
improves the status quo.