Publish holes info via custom LSP notification (#983)
Porting the [`feature/llm`
branch](https://github.com/effekt-lang/effekt/compare/master...feature/llm)
to the current state of the Effekt compiler.

Adds a custom LSP notification `$/effekt/publishHoles` that publishes a
list of holes with type and binder information.

Currently, it doesn't support user-annotated names on holes but `Namer`
generates a fresh identifier for each hole based on the surrounding
top-level declaration, disambiguated by a simple counter. This name
should be unique wrt. to the current module and is somewhat stable to
changes. It breaks when the user changes the order of holes within
definitions of the same name or within the same definition. It also
breaks when the user changes the name of the surrounding definition.