Fix serialization of hole info (#1002)
LSP4j uses the Gson Java library to (de)serialize JSON as part of the
language server protocol. As a Java library, Gson doesn't special-case
the Scala types Option and List (or any other Scala types for that
matter), leading to undesirable output for these based on the default
runtime-reflection based (de)serialization logic.

This PR fixes it by registering custom type adapters for scala `List`
and `Option` with Gson.