Set locale when handrolling JSON (#1018)
IDK why, but on my machine, the output of `--time=json` uses decimal
commas instead decimal points, so `"lexer": { "my/awesome/file.effekt":
14,15 }, ...`
Let's not do that. :)

We should also think about using a proper JSON library, esp. if we use
it in both documentation (#930) and here, it might just be worth to
import
[`uJson`](http://www.lihaoyi.com/post/uJsonfastflexibleandintuitiveJSONforScala.html)
or something...