Deliver holes for literate Effekt files (#1035)
Previously, we only published holes for `.effekt` but not for
`effekt.md`. The reason for this is that for `*.md` files we construct a
new `MarkdownSource` in `compileSource` and we attach the annotations to
that, but we pass the old `StringSource` back to `afterCompilation`,
meaning we don't find the annotations.
This PR fixes this issue by passing the correct source to
`afterCompilation`.