Improve the context info sent for holes (#1029)
This is the first step of improving the information the language server
sends about holes and their context. In particular, we now preserve the
scope that terms originate from, allowing us to display local variables
first.
It does not (yet) order the bindings based on the order they were
defined in because they are internally tracked in `HashMap`s which
unfortunately also means that we apparently cannot show shadowed terms
at the moment?