Natural language holes (#1055)
Introduces holes that can contain arbitrary natural language text. The
syntax is as follows:

```scala
def foo(x: Int): Int = <"
    Holes can splice in Effekt expressions:
    ${ x }
    or statements:
    ${ do bar(); println("hey") }
">
```

The "close hole" code action now only shows if there is a hole with a
single splice (this could be improved in the future, but I consider it
outside the scope of this PR).

Similarly, the extension currently expects to receive a single
(optional) *inner type* for each hole. This PR only provides this if
there is a single splice.

This PR preserves the existing behavior of `<>` and `<{ ... }>`.