Fully resilient lexer (#1066)
Resolves https://github.com/effekt-lang/effekt/issues/522

The lexer is now completely resilient -- should **never** fail.
Instead, it produces an `TokenKind.Error(msg: TokenError)` token which
is currently ignored by the recursive descent parser.

The whole thing is now an iterator (`class Lexer extends
Iterator[Token]`) which simplifies the design a bit.
It also allows us to go much faster, the lexer is now up to 2x faster.

There's also a drive-by in the parser -- we now report all soft fails
(`softFail`) after encountering a fatal failure (`fail`). This means we
can aggregate a few more parse errors when needed.

Additional drive-by: the lexer is now used to highlight code in the REPL
and RecursiveDescent can finally be renamed to Parser

---------

Co-authored-by: Jonathan Brachthäuser <jonathan@b-studios.de>