Add back hack for DeclPrinter (#1086)
In a previous PR, we decided to remove this workaround because we
thought it was likely no longer necessary. It turns out it is necessary,
because `DeclPrinter` contains a partial pattern match, leading to a
`MatchError` when an unexpected symbol is encountered.
Pending a proper fix, this adds back the hack.