Hide document symbols without spans (#1101)
During `ExplicitCapabilities`, we asssign some symbols for the
explicitly passed capabilities (as additional block params).
It might be interesting to add spans to them in the future and maybe
even show them as (optional) inlay hints for educational purposes.
However, in this PR, I just hide them from the list of document symbols
for now, which has previously been cluttered by them.