Try to fix performance regression introduced by a0033bf (#1111)
For whatever reason, before we inlined `BODY` into the else branch

```
val x = if (...) {
  shift(p) { k => () }
} else {
  ...
};
BODY
```

but didn't anymore after a0033bf. 

This PR fixes this by always introducing joinpoints for ifs and matches.
Then, when a continuation is used exactly ones in a follow-up pass, we
will inline it. This way, more things become direct-style, avoiding
pushes. It comes at the cost of potentially not seeing through the
joinpoint (as in the case of `anf.effekt.md`) and of having more known
jumps.

However, our cost model says that known jumps are generally cheaper than
pushes, so it should be worth it.