Output the CPS IR when setting `ir-write-all` (#1151)
When setting `ir-write-all` the CPS IR now also gets written to a file
when used by a compiler (so only the JS compiler right now).

Note that I did not add the CPS case in the `publishIR` method of
`Server` because I am not sure if this breaks something as I don't know
what's going on there.

Related to #935 and #573 and of course only fixes a small portion of the
issue.