Implement namespaced constructors and operations (#1092)
Constructors and operations are now defined in a namespace, named after
the type.

For instance:

```
type TrafficLight { Red(); Green(); Yellow() }
```
defines the constructor `TrafficLight::Red`, but also imports it as
`Red`. The same holds for operations, which can now (in principle) be
disambiguated like `cap.Nondet::flip()`.

As a drive-by, I simplify the way operations are looked up and drop the
unused effect on `Do`, which was in preparation for `do Nondet.flip()`,
which now is `do Nondet::flip()`.

---------

Co-authored-by: dvdvgt <40773635+dvdvgt@users.noreply.github.com>