LLVM: Fix buffering of stdout (#1198)
According to https://stackoverflow.com/a/4201325 , stdout is not
(necessarily) line-buffered when going to a non-interactive output,
which sbt probably is (from the point of view of the compiled Effekt
program). So *all* output in this case is fully buffered.

Here, we explicitly set `stdout` to be line buffered in `main` (using
`setvbuf`).

This fixes #1087 ~(at least for me, but writing an automated test for
this would need changes in the test framework AFAICT)~.
Now includes a test case using filesystem to write to /dev/stdout in
between.

### Alternatives
- Manually `fflush` after every println (see 08f952f / 1e6b928)