Reach reparsability of core pretty-printer (#1181)
Currently, it is very hard to write tests for core. There are two main
reasons for this:

1. The core parser is currently very rudimentary and it is not easy to
write core programs. This would be addressed by a new core frontend
(#1152).
2. The core pretty-printer is currently unable to produce code that can
be re-parsed with the core parser. I would like to address this with
this PR.

While a new core frontend would subsume both of these problems, it is a
lot more effort than adding a re-parsable mode to the current core
pretty-printer. This would allow us to compile source programs to core
and pretty-print them to use them as golden output in core tests.

In particular, when developing the new normalizer (#1127), I would like
to track its behavior in unit tests, but they are currently very
time-consuming to write due to the fact I cannot just use pretty-printed
core for comparison.