Add more streaming combinators (#1244)
These are the utilities I wanted to find in 'stream' (and related
modules) when I was doing AoC'24 and '25:
- `stream::countIf`
- `stream::takeWhile`
- `stream::all`
- `stream::any`
- `list::tails`

Each one has a test and a doc comment.
I tried very hard to add only a single variant of each function so that
we don't have too many single-use functions.
(I also have `countTrue`, `all`, and `any` that work on boolean streams
directly + some variants that work on pull streams)