Parse all definitions even locally, distinguish top-level-only in Namer (#1251)
Resolves #1191

From now on, Namer is to check if a given defn is allowed at top-level.
This also allows us to get a bit further in the pipeline, not everything
has to fail in the parser.
The grammar, and hence the parser, get a bit simpler.

I believe this is morally correct since scoping is the task of the
Namer, not the Parser.