CI: Add macOS LLVM smoke tests (#1256)
There is a small caveat: I only enabled these for the PR CI, not for the
`master` branch CI. If interested in the latter, feel free to request
that too.

---

To be used in #1219.

- run all IO tests on macOS with LLVM
- run with `--debug` flag
- takes some 3m30s (like JS tests)

Here's the log:

<img width="645" height="274" alt="Screenshot 2025-12-13 at 17 50 31"
src="https://github.com/user-attachments/assets/df402a2c-bbb3-400f-94d4-da9ec511418d"
/>