Defer checking of constraints and wellformedness of free variables to speed up compiler (#1277)
We now defer a lot of checks regarding free variables and constraints to
`moduleDecl.typecheck()`. Instead, we construct the free structure of
the free variables computation, like

```
Free.Without(List(x, TInt), Nil,
  Free.Join(..., ...))
```
and have different versions of free variable computations. The probably
cheapest one is `freeIds` as it really just computes a set of ids. The
next cheapest one is `freeValues` and `freeBlocks` which eagerly
overwrite and subtract ids without checking them. Only when calling
`Free.wellformed()` it is checked that the types of free variables work
out.

The results are cached so that we don't recompute them over and over
(except the checking).

Additionally, the call to `moduleDecl.typecheck()` is hidden behind a
debug flag. So wellformedness will only be checked whenever effekt is
called with `--debug` or you insert calls manually during development.